Register TaskbarCreated message to reshow tray icon
If the explorer process is restarted, tray icons need to be readded,
otherwise they wont show up. The shell registers TaskbarCreated as a
message and then broadcasts it to all top-level windows when the taskbar has
been created. When this message is received, the tray icon needs to be
readded.